// src/app/services/models/country.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MobileMoneyOption } from '../models/mobile-money-option';
export interface Country {
  acronym?: string;
  countryId?: number;
  currency?: string;
  mobileMoneyOption?: Array<MobileMoneyOption>;
  name?: string;
  rating?: number;
  region?: string;
}